requirements.txt
clinner==1.12.3
colorlog==3.2.0
Django==2.2.13
djangorestframework==3.11.0
gitdb==4.0.5
GitPython==3.1.3
health-check==3.4.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
